refactor(searchcore): move tokenizer into module (P1.2)
