+++ internal/core/symbols/function.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/codetrek/haystack/internal/conf"
 	"github.com/codetrek/haystack/internal/core/invertedindex"
-	"github.com/codetrek/haystack/internal/core/invertedindex/tokenizer"
+	"github.com/codetrek/haystack/searchcore/tokenizer"
 	"github.com/codetrek/haystack/internal/core/pebble"
 )
