+++ internal/server/searcher/searcher_cjk_test.go
@@ -4,7 +4,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/codetrek/haystack/internal/core/invertedindex/tokenizer"
+	"github.com/codetrek/haystack/searchcore/tokenizer"
 	"github.com/stretchr/testify/assert"
 )
