+++ searchcore/tokenizer/BENCHMARK.md
