+++ searchcore/tokenizer/ascii_tokenizer.go
