+++ searchcore/tokenizer/benchmark_test.go
