+++ searchcore/tokenizer/cjk_detect.go
