+++ searchcore/tokenizer/cjk_detect_test.go
