+++ searchcore/tokenizer/cjk_tokenizer.go
