+++ searchcore/tokenizer/cjk_tokenizer_test.go
