+++ searchcore/tokenizer/mixed_tokenizer_test.go
