+++ searchcore/tokenizer/tokenize.go
