+++ searchcore/tokenizer/tokenize_for_index_test.go
