+++ searchcore/tokenizer/tokenize_for_search_test.go
