+++ searchcore/tokenizer/tokenizer.go
